requirements.txt
Django==1.11.6
djangorestframework==3.10.0
psycopg2==2.8.6
python-decouple==3.4